start organizing code into modules for easier readability and maintenance
