+++ tadpole.py
@@ -3,6 +3,7 @@ from PyQt5.QtWidgets import *
 from PyQt5.QtGui import *
 from PyQt5.QtMultimedia import QSoundEffect
 from PyQt5.QtCore import Qt, QTimer, QUrl
+from tadpole.ui.dialogs import MusicConfirmDialog, ReadmeDialog
 
 # OS imports - these should probably be moved somewhere else
 import os
@@ -305,168 +306,15 @@ class BootConfirmDialog(QDialog):
         # Load Initial Image
         self.current_viewer.load_from_bios(self.drive)
 
-class ReadmeDialog(QMainWindow):
-    """
-    Dialog used to display README.md file from program root.
-    """
-    def __init__(self):
-        super().__init__()
-
-        self.setWindowTitle("Read Me")
-        self.setWindowIcon(self.style().standardIcon(QStyle.StandardPixmap.SP_TitleBarContextHelpButton))
-
-        self.text_edit = QTextEdit(self)
-        self.text_edit.setReadOnly(True)
-        self.text_edit.setMinimumSize(500, 500)
-        self.setCentralWidget(self.text_edit)
-        try:
-            with open(os.path.join(basedir, "README.md"), "r") as readme_file:
-                self.text_edit.setMarkdown(readme_file.read())
-        except FileNotFoundError:  # gracefully fail if file not present
-            self.text_edit.setText(f"Unable to locate README.md file in program root folder {basedir}.")
-
-
-class MusicConfirmDialog(QDialog):
-    """Dialog used to confirm or load music selection with the ability to preview selection by listening to the music.
-    If neither music_name nor music_url are provided, allows import from local file.
-
-        Args:
-            music_name (str) : Name of the music file; used only to show name in dialog
-            music_url (str) : URL to a raw music file; should be formatted for use on SF2000 (raw signed 16-bit PCM,
-                mono, little-endian, 22050 hz)
-    """
-    def __init__(self, music_name: str = "", music_url: str = ""):
-        super().__init__()
-
-        # Save Arguments
-        self.music_name = music_name
-        self.music_url = music_url
-        self.music_file = ""  # used to store filename for local files
-
-        # Configure Window
-        self.setWindowTitle("Change Background Music")
-        self.setWindowIcon(QIcon(self.style().standardIcon(QStyle.StandardPixmap.SP_MediaVolume)))
-        self.sound = QSoundEffect(self)  # Used to Play Music File
-
-        # Setup Main Layout
-        self.layout_main = QVBoxLayout()
-        self.setLayout(self.layout_main)
-
-        # Main Text
-        self.label_confirm = QLabel("<h3>Change Background Music</h3><a href='#'>Select File</a>", self)
-        if self.music_name == "" and self.music_url == "":
-            self.label_confirm.linkActivated.connect(self.load_from_file)
-            pass
-        self.label_confirm.setAlignment(Qt.AlignmentFlag.AlignHCenter)
-
-        self.layout_main.addWidget(self.label_confirm)
-
-        # Music Preview Button
-        self.button_play = QPushButton(QIcon(self.style().standardIcon(QStyle.StandardPixmap.SP_MediaPlay)),
-                                       " Preview",
-                                       self)
-        self.button_play.setDisabled(True)  # disable by default
-
-        self.layout_main.addWidget(self.button_play)
-        self.button_play.clicked.connect(self.toggle_audio)
-
-        # Main Buttons Layout (Save/Cancel)
-        self.layout_buttons = QHBoxLayout()
-        self.layout_main.addLayout(self.layout_buttons)
-
-        # Save Button
-        self.button_save = QPushButton("Save")
-        self.button_save.setDisabled(True)  # disable by default
-        self.button_save.clicked.connect(self.accept)
-        self.layout_buttons.addWidget(self.button_save)
-
-        # Cancel Button
-        self.button_cancel = QPushButton("Cancel")
-        self.button_cancel.clicked.connect(self.reject)
-        self.layout_buttons.addWidget(self.button_cancel)
-
-        if music_name and music_url:  # enable features only if using preset options
-            self.label_confirm.setText("<h3>Change Background Music</h3><em>{}</em>".format(self.music_name))
-            self.button_save.setEnabled(True)
-            self.button_play.setEnabled(True)
-
-    def toggle_audio(self) -> bool:
-        """toggles music preview on or off
-
-            Returns:
-                bool: True if file is playing; false if not
-        """
-        if self.sound.isPlaying():
-            self.sound.stop()
-            self.button_play.setIcon(QIcon(self.style().standardIcon(QStyle.StandardPixmap.SP_MediaPlay)))
-            self.button_play.setText(" Preview")
-            return False
-
-        else:
-            if not self.sound.source().path():  # fetch and convert raw file if not already done
-                self.button_play.setDisabled(True)  # disable button while processing
-                file_data = self.get_and_format_music_file()
-                if file_data[0]:  # fetch/conversion succeeds
-                    self.sound.setSource(QUrl.fromLocalFile(file_data[1]))
-                    self.button_play.setEnabled(True)
-                    self.button_save.setEnabled(True)  # enable saving as well since file seems OK
-                else:  # fetch/conversion fails
-                    self.button_play.setText(file_data[1])
-                    self.button_play.setIcon(QIcon(self.style().standardIcon(QStyle.StandardPixmap.SP_MessageBoxCritical)))
-                    return False
-
-            # format button and play
-            self.button_play.setIcon(QIcon(self.style().standardIcon(QStyle.StandardPixmap.SP_MediaStop)))
-            self.button_play.setText(" Stop")
-            self.sound.setLoopCount(1000)
-            self.sound.play()
-            return True
-
-    def get_and_format_music_file(self) -> (bool, str):
-        """Downloads/loads and re-formats the raw music file as wav.
-
-            Returns:
-                tuple (bool, str): True or false based on success in fetching/converting file, and path to resulting
-                    temporary wav file, error message if failed.
-        """
-        if self.music_url:  # handle internet downloads
-            try:
-                r = requests.get(self.music_url)
-                if r.status_code == 200:  # download succeeds
-                    raw_data = BytesIO(r.content)  # read raw file into memory
-            except requests.exceptions.RequestException:  # catches exceptions for multiple reasons
-                return False, "Download Failed"
-        else:  # handle local files
-            with open(self.music_file, "rb") as mf:
-                raw_data = BytesIO(mf.read())
-
-        wav_filename = os.path.join(basedir, "preview.wav")
-        with wave.open(wav_filename, "wb") as wav_file:
-            wav_file.setparams((1, 2, 22050, 0, 'NONE', 'NONE'))
-            wav_file.writeframes(raw_data.read())
-            if wav_file.getnframes() > (22050*90):  # check that file length does not exceed 90 seconds (max for Froggy)
-                return False, "Duration Too Long (90s max)"
-        return True, wav_filename
-
-    def load_from_file(self) -> bool:
-        file_name = QFileDialog.getOpenFileName(self, 'Open file', '',
-                                                "Raw Signed 16-bit PCM - Mono, Little-Endian, 22050hz (*.*)")[0]
-        if file_name:
-            self.music_file = file_name
-            self.music_name = os.path.split(file_name)[-1]
-            self.label_confirm.setText("<h3>Change Background Music</h3><em>{}</em>".format(self.music_name))
-            self.button_play.setEnabled(True)
-            self.toggle_audio()
-            return True
-        return False
-
 
 # SubClass QMainWindow to create a Tadpole general interface
 class MainWindow (QMainWindow):
-    def __init__(self):
+    def __init__(self, base_dir):
         super().__init__()
+        self.base_dir = base_dir
+
         self.setWindowTitle("Tadpole - SF2000 Tool")
-        self.setWindowIcon(QIcon(os.path.join(basedir, 'frog.ico')))
+        self.setWindowIcon(QIcon(os.path.join(base_dir, 'frog.ico')))
 
         widget = QWidget()
         self.setCentralWidget(widget)
@@ -522,7 +370,7 @@ class MainWindow (QMainWindow):
         self.tbl_gamelist.cellClicked.connect(catchTableCellClicked)
         layout.addWidget(self.tbl_gamelist)
 
-        self.readme_dialog = ReadmeDialog()
+        self.readme_dialog = ReadmeDialog(self)
 
         # Reload Drives Timer
         self.timer = QTimer()
@@ -599,10 +447,10 @@ class MainWindow (QMainWindow):
     def change_background_music(self):
         """event to change background music"""
         if self.sender().text() == "From Local File...":  # handle local file option
-            d = MusicConfirmDialog()
+            d = MusicConfirmDialog(self)
             local = True
         else:  # handle preset options
-            d = MusicConfirmDialog(self.sender().text(), self.music_options[self.sender().text()])
+            d = MusicConfirmDialog(self, self.sender().text(), self.music_options[self.sender().text()])
             local = False
         if d.exec():
             if local:
@@ -932,8 +780,13 @@ if __name__ == "__main__":
     # Initialise the Application
     app = QApplication(sys.argv)
 
-    # Build the Window
-    window = MainWindow()
+    # Get the working directory and build the main window
+    if getattr(sys, 'frozen', False):
+        # For determining path after program is built with PyInstaller to single file
+        application_path = os.path.dirname(sys.executable)
+    else:
+        application_path = os.path.dirname(os.path.abspath(__file__))
+    window = MainWindow(application_path)
 
     # Update list of drives
     window.combobox_drive.addItem(QIcon(), static_NoDrives, static_NoDrives)
